+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `DataCube.resample_spatial()` now supports parameterized `resolution` and `projection` arguments.  ([#897](https://github.com/Open-EO/openeo-python-client/issues/897))
 - Sanitize asset download filenames (e.g. strip slashes, (semi)colon, hash, ...), instead of blindly using the asset key as filename. ([#820](https://github.com/Open-EO/openeo-python-client/issues/820))
 - Support parameters in `DataCube` apply- and band-math operations ([#903](https://github.com/Open-EO/openeo-python-client/issues/903))
-- Introduction of simple (experimental) event system at `Connection.events` to allow hooking into various events of the API and internals
+- Introduction of simple (experimental) event system at `Connection.events` to allow hooking into various events of the API and internals ([#922](https://github.com/Open-EO/openeo-python-client/issues/922))
+- Initial implementation of pytest plugin `auto_list_job_ids` to automatically list created jobs and sync processing requests ([#922](https://github.com/Open-EO/openeo-python-client/issues/922))
 
 ### Changed